- fix: due to an overflowing short, the blacklist timer would eat 100% cpu  for 2 ticks (~125ms) each minute, after 64 min. running time  (as opposed to unstable the too much time spent in cleanup check will fire   and exit the function after 2 ticks, but still you'll end up with degraded performance)

+++ dst_blacklist.c
@@ -275,8 +275,8 @@ inline static int _dst_blacklist_clean_expired_unsafe(unsigned int target,
 												ticks_t delta,
 												ticks_t timeout)
 {
-	static unsigned short start=0;
-	unsigned short h;
+	static unsigned int start=0;
+	unsigned int h;
 	struct dst_blst_entry** crt;
 	struct dst_blst_entry** tmp;
 	struct dst_blst_entry* e;
